getAllTagsOrderByName

+++ src/main/java/com/javamentor/qa/platform/models/dto/TagDtoPagination.java
@@ -16,5 +16,5 @@ public class TagDtoPagination {
     private String name;
     private String description;
     private LocalDateTime persistDateTime;
-    private int questionsCount;
+    private Long questionsCount;
 }
